Implemented Procedural Mesh generation
A "Floor" object has been created with a script that does procedural mesh generation of a cube. It gives the user the option to change the size of the "Floor" object, which is designated to be a cube. The user is also able to apply a positional offset to the texture on each face of the cube to help with allignment.