Modified Detection method for portal teleporting
The method for detecting if the player will collide with a portal teleport trigger is now done during each moment the player will be moved. It looks ahead for a teleport trigger and runs it's TeleportCollider function if needed. If no collision is met, it will let the player move on.