Dynamic camera rect calculation using camera's viewport and portal's vertices
The CalculateViewingRect() function has been upgraded so that now the camera will take into account the fact that some vertices of the mesh will be outside the camera's viewport.

It does this by calculating the edges of the mesh and where they connect to the edges of the camera's view.

HOWEVER, THIS REQUIRES THAT THE PORTAL MESH IS CREATED USING PortalSet's CreateDefaultMesh() function.