Implemented Layering system for double sided portals
Portals can now be set to be double sided, which makes both their normal and backwards portals enabled. It will alos apply a rendering layer that makes each portal ignore it's reversed portal.

When a portal is not double sided, a boolean value can be checked to enabled either only the normal or the backwards portals.

However, when a portalSet has an offset, the player will still be able to see the backwards portal if they are really close inside the portal. This will be fixed with the implementation of invisiblePortals for the player.