PuzzleRoom creation adjusted
The puzzleRoom handles creation better and now includes a collider covering the entire room. This allow the clouds on the top and bottom edges to track the player. Also implemented is the player will teleport from the top to the bottom and vice versa if they fall too far from the main puzzle area.