Added a footstep tracker
Track the previous series of footsteps to calculate the average direction the player was heading in. This allows us to detect if the player is attempting to undergo a large momentum change, like a sharp turn or a 180.

Also added a basic footstep sound along with a play area for testing.