Expanded footstep calculation
Added a more in-depth calculation to the footsteps by observing the previous tracked inputs and comparing their average direction with the player's current inputted direction. The same goes for the spped/magnitude of the directions.

This makes the footsteps seem more dynamic, with the inclusion of quick side to side steps, sudden stops, sudden speed up/down and even sharp turns having an effect on when the next footstep sound will occur.