Refactored footstep calls
The way a footstep is requested to be played has been changed to allow a muhc easier time to request to play a footstep effect.

Opens up the Landing sound options by having soft, normal and hard landing functions. A player will play a sound when they "Land", which is the act of going from an airborn state to a grounded state.