Implemented muted and upgraded audioSources
There are now two music sources, one plays the "muted" clip and the other plays the "upgraded" clip. It is planned for the music to switch to the upgraded version when the player completes a room and for the music to go back to a muted version when they fall off the level.

There is also sepperate fade values for both sources along with a function that swaps between the two music sources, SetMusicFade().