Added further box and circular mesh customizability
Boxes now have a seperate width for the top and bottom parts of the box.

Circular filler meshes can now have their radi set. The user also has control over how far and when the sine function starts in relation to the maxBumpWidth applied to each vertex of the cylinder.