Pillar's center column now uses the seed to generate it's stats
The center column's radius is now set using the seed. It is also set to ensure the center column will never go past it's self-impossed limit of the baseWidth. It even takes into account the bump radius after the stretch and offset, which have been changed to go from the range of [0, PI*2] instead of [0, 1].