Dynamically set the amount of vertexes used to define a circle in a circular mesh
When creating the mesh for a circular mesh, the radius of the circle will be used to determine how many vertices will be used when defining the mesh. When handling a circular mesh with a cruve that uses a series of vertex points, take the largest radius of the set of points and use it as the reference radius.