Filler objects can now base their width on the center column's end width
When creating filler objects, we can now have their width be based on the end of the center column's radius. IE, instead of being half of the baseWidth, it can now be half of the top of the center column's radius.

This allows both bottom and top fillers to remain the same look while accomadating to the different radius of the top and bottom of the center cylinder.