Filler radius is now relative to the pillar's radius
When defining filler objects, the radius will now be relative to the baseWidth and the main column's radius. This is done by having the filler radius use a range of [1, 0] , where 1 is 100% the baseWidth while 0.25 is 75% the main column radius.

This is also true for the circular meshe's bump radius - the value of the bump is the radius of the mesh at the apex of the bump, which now lets us have bumps bend inward.