When creating filler, use a width that is relative to the remaining height
Two expectedFillerWidth values are used when creating filler. They are set from the range of [1, 0] and are relative to how much remaining height in the filler is left. The goal of these values is to ensure each filler object will form a frustum from the base to the main cylinder.

Square fillers have been modified to use these values while circular filler will require more effort to implement.