Basic implementation of a new "stairs" filler and adjusted square filler UVs
A new filler object of "stairs" has been added that features a series of square filler as they shrink in width. It should be used when there is not a lot of height but a lot of width.

The adjustement to the UVs is because the "stairs" has many squares on top of eachother and the top and bottom of the square UVs look better using this new version.