Fixed how filler creation handles radians and added a new filler object
Before, the filler creation process did not handle the startingRad and radInc values correctly when creating a circular filler object. It now is currently seperated into two for loops depending on whether the filler being created is above or bellow the main column.

A new filler object has been created, but it's parameters have not yet been set. It aims to cover a lot of height and width using one large circular using only a quarter of a sine wave. It is sandwiched between two other circulars.