Deciding which filler object to use is now random
As long as each condition on a filler object is met, when deciding what filler object to create, pick a random filler object that meets their required conditions. It is currently done in a inefficient way, so a better version should be done. But currently, filler object creation randomnly chooses a filler from a list of objects that meet the requirements.