Refactored how the pillar handles it's individual pieces. It is now split into seperate objects
When the pillar is creating itself, each piece of the pillar is now a seperate object that is added as a child to a container, which is a child to the main pillar object which holds the script.

This allows each piece to be more independent of eachother and the pillar is now not one singular mesh. Each collider will also be a part of their respective object.