Small Cicrular pillar objects now use multiple box colliders. Seperated child pillar piece object creation into a single function.
When a circular mesh is shorter than it is is wide, it will now use multiple box colliders, which are each a unique child of the circular mesh object.

When a piece of the pillar is being created, it's gameObject along with it's colliders are created using a seperate function to make the functions more readable.