Adjusted The TestArea's transforms and objects to be much more readable and accepting of other areas
A more general layout for a puzzle room has been applied to the testArea. Also, the testArea's first puzzle is now "complete" for the time being.

Connecting rooms have their own container for now, but if rooms decide to de-load if not inside them, then each connector might need to be a part of the room for convenience.