Started transitioning player movement into Fixedupdate
All player movement will be changed to be done in Fixedupdate and will use Rigidbody.MovePosition. This will prevent the player from clipping into walls and objects when moving around.

However, this will break how the player detects themselves connecting to a portal, so it will need to be re-implemented. So far, the best idea has been to save the player's current postition at the end of an Update call, and look back to it once the next Update call is made and check if any portal is between the two positions.

Due to the order of the functions (FixedUpdate > Update > SceneRender), this should be able to properly teleport the player seamlessly just like before.

Any potential problems occur when the player teleports into an object due to the teleportation happening at the Update function.