Moved movement into FixedUpdate and commented in the new movement implementation
The player movement is now done in fixedupdate. The way the player moves using MovePlayer(vector3) is now changed as it should only send a request ed direction, not a teleport. Teleporting will be added into a new function that will handle teleporting AFTER the player has moved.

This means that for a player to pass through a teleporter, they must be able to fit in said space. IE, if there is a wall behind a teleporter, the player will not be able to teleport if the wall blocks the player's center from passing the teleporter's trigger plane.

Also, I do not expect the player to be able to teleport currently. Yet, they still can.