Player steps are now fixed
Player steps are now properly updated at a consistent rate and the act of calling the step tracker has been moved into  the fixedUpdate function. It still functions as normal but it requires an extra varaible to track any extra movement that may have been caught in the Update function.