WaitingRoom now has a trigger for it and it's AttachedRooms
A WaitingRoom now creates a trigger that encompasses itself and most of the two AttachedRooms connected to it. The creation of this trigger is fairly convoluted and having each AttachedRoom save certain of it's stats may help in the long run.

Also, the CustomPlayerController now properly ignores stray trigger colliders it may encounter.