Fixed the issue of the player's camera seeing through objects above when jumping
When the player jumps and collides with a flat object above, the player's camera ends up sitting directly on the plane as the ray trace will not move backwards but simply stops at the collision point.

By using two ray traces, one to scout ahead and see how far the camera can reach, and another to fire the actual ray, but this time having less distance than the scout, we can prevent this scenario from occuring.

However, the situation of the rayTrace reaching very close to a parralel plane will still have the camera be very close to the wall. This situtation is a rare scenario as the player's camera only ever leaves their body's collider if they are fast falling.