Fixed Window's portal rendering. Portals will now ignore very small views when rendering.
The portal's collider mesh used to detect how much of the portal the camera was viewing was not properly updated after the window set it's new sizes.

Also, portals will not not render their view if the camera cannot see enough of the portal. The amount of portal that is required to see before it renders is very small and will not effect the user's experience. This implementation is to prevent useless renders and to prevent trying to render using a rect with a viewing size of 0. This does fix the random matrix4x4 errors that were occuring to the portals during run-time.