Added a spherebox thats placed around the outside window of a WaitingRoom
The outside windows of a WaitingRoom now have a sky sphere surrounding them to simulate a new environment. The size of this sphere and where it will be placed is done by the player, but it should be procedural for safety. The Material used by the skySphere must be given to the WaitingRoom.

Maybe for the best effect, add extra windows in the WaitingRoom for more views. Also, the sky sphere should move along with the player's camera for it to seem like a real skybox.