Fixed WaitingRoom's collider. Window's skysphere now moves with the player's camera
The WaitingRoom's trigger that emcompases the WaitingRoom and its connected AttachedRoom is now properly being set.

Whenever the player is inside the WaitingRoom's trigger, adjust the position of the sky sphere of the outside window to reflect the player's movement. This is to make the sky sphere seem more like a proper skysphere instead of a small sphere placed around the window.