Stairs now use two sets of two points to define it's stairs
A set of stairs requires 4 points now. This way the top and bottom of a set of stairs can have different widths, or maybe even sepperate heights on either side. This makes stairs more dynamic and easier to create by having a plane define the stairs at the start. Rotation/direction of the stairs is not yet determined, but it will be implemented soon.