A plane is now created to represent the stairs
The stairs can now be seen as a plane. The set of stairs has been changed to use 3 points to define it's position. No matter the position of the start and end points, no matter the rotation of the "sideAngle", there is always a plane in a legal position.

Next is to use a "stairsAngle" to add the steps onto the plane, making it a real set of stairs.