Added side parts to the stairs model and UVs.
Side part of the stairs are now modeled. The UVs have been applied, but are not perfect and require a fix. Possibly two materials for both directions of the steps?

Removed the "bellowPlane" variable as the same effect can be done by switching the start and end points and to create the mesh for the sides of the stairs will require another creation function.