MeshCollider added to the stairs
The stairs will now create it's own meshCollider. This mesh collider is accurate with the base, but the steps are too precise and would require lots of box meshes. Whether it is more efficent using a smaller meshCollider + a bunch of box colliders for the steps I do not know. But for now, the stairs will have a slope to represent the steps of the stairs.