Added a second material for the upward steps. Fixed the UVs of the steps
The stairs now requires two materials - one for the base and forward steps and another for the upward steps. This is to give a clear depth to the steps that can be seen from most angles. Now the only thing left is to fix the UVs of the step's side and base.