The Stair's transform can now be moved to properly position the stairs
Before, the stairs were not placed in the correct spot if the transform with the script was moved. Now, we can move the stairs around a puzzle room and it will be placed in the correct spot.

Also, created some empty rooms to have an easier time creating new rooms.