Removed "Stop player upward momentum".
When the player jumps upwards and is travelling againts a ceiling continuously, it doesn't feel good for it to stop as the jump ends prematurly. So, we always want the jump to be the same every time.

Also, adjusted the puzzle room 3's upside down ceiling of the entrance way so that when the player is oriented the correct way, the jump to the exit is not as difficult.