CubeCreator can now use two materials
The CubeCreator script can now use two materials. By having a list of booleans representing each face of the box, we can select which boxes to use which material. Each boxCreator will need to track it's material through it's function now, which has been implemented to all CubeCreators already in the scene.