Hacked together a fix for the teleport-step bug
There is a bug currently when travelling between two portals, with one being at an off angle, in this example the portals in puzzle room 4.

When stepping through the portal, the player will undergo a step, possibly from a past movement request that is told to occur before a teleport, then is accepted and run after the teleport. There is a hack together to fix it, indicating the problem occurs within the teleport frame.