CubeCreator now updates upon script value change. Stairs now has a "Up" vector value to change the flat angle.
The cubeCreator now recreates the cube if any values of the script changes.

The stairs script can now change what is considered it's "up" vector so that when it wants to reset the stairs angle to have them flat, it will use the given upVector to determine what is considered up.

Also added a few more stairs to puzzle room 4.