Added a bunch of content in puzzle room 4
There is now an actual puzzle in the puzzle room 4. The portals need to be checked to see if they are bugged or not. They can be fixed by mesing with their rotation.

Also, a final portal needs to be added as a proper exit. The goal should be to reach the exit upside down, ie walking in the roof of the next room.