Fixed and changed a few quirks in the game. Description for more details.
The player will only enter FastFalling state if they are falling past a certain speed.

The player camera's Yoffset will be reduced faster depending on how close to it's limits it is.

Fixed the WaitingRoom's windows not properly being created and the portla not having a proper offset.

Fixed the window's borders not being created when the windowScript called it's CreateCube function.