Window sizes are relative to the room's sizes. SkySpheres are placed
The sizes of the windows for each WaitingRoom is now determined by a given ratio of how much of the wall it will occupy. Also, the sky spheres are placed by the script relative to the waitingRoom's positions soi that the portals do not attempt to render the puzzle room portals. The size of the sky spheres have been increased, but they still do not collide with eachother