TerrainController now creates and destroys terrain relative to a given position
The main controller now requires a given position to use as a point of reference. The goal is to render the terrain around the given position. It is currently working but it is not efficiant as it continuously deletes and recreates the same chunk.

Also, the scripts have been cleaned up to be mor readable.