Terrain creation/removal much more efficent
The Updating of the terrain has been made much more simple using the Except() and ToList() functions from System.Linq. It let's us generate a list of all loaded chunks, then extract all chunks that need to be removed and new ones that need to be loaded. This makes the terrain updating much more efficient.