Chunk Creation/Generation has been added to the ChunkCache. Creation is done using a Dictionary as a queue
When wanting to add a new chunk to the terrain, a request to add a chunk is sent to the cache, where it adds a chunk object (without it's generated terrain) and it's key to a ChunksBeingGenerated collection. This collection is scanned every frame and the chunks inside it are generated, then added to the LoadedChunks dictionary.

When removing chunks, it can encounter the request to remove a chunk that has not yet been generated. In this case, the chunk is removed from the ChunksBeingGenerated before reaching the loadedChunks dictionary.