The TerrainController now follows a given object. The StartingRoom now links the player to the terraincontroller once they leave after the glass breaks
When the player breaks the glass in the startingRoom and then leaves, the startingRoom script will link the player's transform to the terrainController. This will make the terrainController observe where the player is located and generate chunks around them.

A problem that can occur is that the player breaks the glass, then leaves the room NOT from the window. This should easily not occur assuming the player is at the expected 90 degree angle and not some odd 45 degrees. Either way, we do not want the player to leave the room from any way other than the window portal once the glass breaks.