Added a Path to the noiseProvider through implementing an extra layer to the noise
The noise provider now uses a layer of a "path" that will override the current noise function. This is a basic implementation and further improvements will be made, but for now it is expected that a layer will have a ratio and a height map. The path is currently implemented into the current system.