Implemented the biomes into the system
The biomes can now control the stats of an area in the noise function, such as max noise hieght.  We can control how often biomes change with the biomeFrequency value.

To add some realism to the map, a second noise function should be used. This second noise function will serve the same purpose of octaves and should be added to after the noise is calculated in the GetBiomeNoise.