Added new textures for the terrain
The terrain is now using proper textures that are applied by the biomes. The slope of the terrain does not control how it looks anymore, but the code that uses it still exists as a comment. This code may have some reuse as using a single texture for a whole biome is not ideal.