Implemented Steepness back into the terrain texturing process
When the terrain is being textured, it will use two textures for each biome, with the steepness of the terrain determining how much of each texture is used. The steepness's max angle should be set for each biome, but for now is the same for each. Currently, by using a white texture as each biome's maxAngle, it looks like it is currently winter, especially near the moutains.