Fixed a bug where the game would lock up when the player fell out of the StartingRoom
When the player was falling out of the startingRoom, sometimes their leg's rayTrace would hit the room's trigger and get stuck in a loop.

This is fixed by having the window's portal's trigger's thickness not be set to 0 as having rays pass through a portal with a 0 thickness is not ideal.

Also, the startingRoom's trigger was not properly set to the IgnoreRaycast layer.