Adjusted how the sky sphere is handled along with clipping planes
The player camera's far clipping plane controls how large the outside's sky sphere is along with other values, such as the waitingRoom's outside window position and the portal camera's far clipping plane.

An issue that occurs involves objects that are near the far clipping plane. When falling outside, the terrain that is really far away has rendering issues. This can hopefully be resolved by implementing fog or some kind of visual effect on terrain that is very far away. Also, LoD on the terrain should be controlled through scripts as teleporting from the window to outside has a wierd effect on the terrain as it does not look similar from the player's camera and the portal camera.