Added fog and fixed some rendering issuess with the terrain
A fog effect has been added to the world. It makes far away terrain look better, but the skysphere will need to be changed as we do not want it to be effected by the fog.

Also, the flickering textures of far away terrain has been fixed by setting the terrain's base map distance to reflect the player camera's far clipping plane.

The terrain's edges flickering when connecting to the sky sphere is fixed by increasing the player's near clipping plane. The player's near clipping plane was very small due to the use of portals, but now we wont be using more portals so it's safe to increase it.