Adjusted how the skySphere is rendered for the player camera
The SkySphere is now rendered through a different camera for the player. By adding a script to the object that holds a camera, we can make that camera ignore the skybox and have another camera as a child that will render the skySphere. This new camera can then have sepperate effects, such as ignore the fog.