The SkySphere camera now ignores fog
When a SkySphereCamera is created through it's script, it will attach an IgnoreFog script to the camera, which will cause the camera attached to the object with the script to remove any fog effects before rendering it's scene, then reapplying them after it's done. This allows the skysphere to be rendered behind the fog.