SkySphere is now immune to fog and cleaned up the previous fog ignoring methods
Now that we have a shader that ignores fog for the materail it uses, we do not need the previous scripts that altered the camera's rendering, so we returned the player camera's rendering stats back to normal. The SkySphere layer is not required anymore, but we will still keep it for now.

Also, the fog's distance is set by the terrainController and it uses the player's farClippingPlane as reference, just like the skySphere.