The Terrain textures are now set during a thread
When the terrain is being generated, the terrain's hieghtMap and textures are done in a thread, as both of them require access to the noiseProvider.

Also increased the player's render distance and increased the chunk's length. This allows the player to see further and keeps the same resolution of the terrain.