WaitingRoom's windows now only render it's skySphere and outside window
When the waitingRoom creates it's portals and windows, it will only allow it's window's portal's cameras to render only the skySphere and it's window frame. This will prevent any unnescecairy rendering of outside objects such as the terrain or puzzle rooms.

However, it will still render the windows of the other waitingRooms if they are within view.