Terrain can now be generated from threads
When the focus point cahnges and the TerrainController starts generating terrain using threads, the terrain texture now uses a thread to generate the splatMap ratio due to the biome.

Later, upon the terrain's object creation, it will do another pass over the terrain to apply the steepness of the height to the splatmaps.