Attempting shaders on terrain
I am trying to use shaders with the terrain in an attempt to fix the distance issue when handling textures.

No matter the shader and material, I am encountering issues involving the splatmap of the terrain. Perhaps to futher test it, the splatmap can be lowered to only use one texture and see the effects.