The terrain is now in it's final form
I was unable to apply a shader or fix the distance issue involving the terrain so I will leave it as it after a few changes to the pixelError.

Other optimizations may come into play, such as when loading a new chunk, but for now it's visuals are staying as is.