Created timings for the procedural terrain creation
By timing the creations of the terrain, we can find out where most of the time is spent when running functions. We can see that the terrainController takes around 60ms when generating a new terrain chunk. The actual chunk generation can be split to show where most of the time is spent, which is shown the steepeness function, which is used to add depth to the two textures used on each biome.