Terrain is now forced to load in startup and uses a coroutine during runtime
Now when terrain is generated during runtime, the chunk will start a coroutine once it's terrainMap is ready and is moved into the loadedChunks dictionairy in the chunkCache. At this point, it is safe to fix the terrain's splatMap and re-flush the textures once it's coroutine is finished.

We may need to implement another dictionary in the ChunkCache for any chunks that are using a coroutine to fix it's textures. This is because if we have to many coroutines running, it will still cause a frame loss.