Trying to get the terrainController to clone a default terrain
Most of the terrain's overhead is in it's Constructor now. Since this overhead is universal for each terrain object, we could simply run it once and clone the object multiple times whenever we need a new terrain.

This is the idea, but for some reason due to serialization i cannot properly Instantiate a terrain object while keeping certain variables, such as the splatMap. This is confusing as if I were to simply duplicate a terrain object that has already been fully created, it will already have the same splatMap and textures.

We just need to find a way to clone a pre-existant terrain with textures and all, then modify it's heightMap.