The terrainChunk is now instantiating a previous terrain object
The act of instantiating a previous terrain object is shown to also take around 15ms of time. This brings up a new problem where creating any kind of terrain chunk will always suffer from the overhead cost of creating a new terrain object.

A potential fix for this would be to use object pooling so that when a chunk is "removed", it will simply reset and will be used for the next chunk that needs to be loaded. That way the terrainData is already created and assigned it's textures maps and the only changes is the heightMap and the textureMap, both if which are done through threading and coroutines.