Chunk creation is now done using the inactive chunks in the pool
When a new chunk is added into the game, it uses an inactive chunk from the pool of pre-created chunks. When chunks are removed, they are sent back to the pool and can be taken out and used again.  However, the removal and reintergration process is not working properly as the new chunks being added are using the wrong terrain/coordinate. Also, some stray chunks are not being removed/added.