Added a script to update the shader's values
The shader now has two float values that control when and for how long it takes the texture to blend. It can even use a focus object to alter where the blending begins. The shader is comming along nicely and in the testScene, it seems as if the focus point serves as a light.