The terrain textures now highlight the relative steepness of each biome
Each biome now uses the same texture as each other, where one terrian texture is full white and the other is full black. This now highlights the terrain's relative steepness, and shows how there is a wierd where the steepness that is calculated on the edges of a terrain are incorrect.

Upon further inspection, the negative X and negative Z edges are incorrect.