terrainData.GetSteepness does not return proper values at the edge cases
Currently, when calling GetSteepness, when either given points are at the edge of the terrain (0 or 1), then the steepness returned will not be ideal. This issue might be because the neighboor is not set? Anyway, a hack has been implemented for the terrain to take the neighbooring point on the same terrain tile instead. It doesn't seem good, but it looks better.