Cannot find a fix for the steepness issue, current state will be the final version
currently, the steepness when on the edge case will use the steepness of the adjacent position on the same terrain. This is no ideal as there is a clear cut between certain positions where the steepness changes drasticly. However, a fix cannot be fund so it will have to suffice.