Created a basic GlobalRoomController, which aims to control room creation
A GlobalRoomController script is created and attached to some empty object in the scene. It requires links to the containers holding the main rooms of the game.

So far, it can repopulate it's arrays using the rooms laready created and it can rename the rooms. It is important that each room is in the proper position in the hierarchy as their index in the child structure is important.