Can now move puzzleRooms in 3D using GlobalRoomController
The user can now input values into the globalRoomController to be able to move any puzzle room. Any changes in it's height will move it's entrance and exit positions, which control the positions of the linked attachedRooms.

By changing the attachedRoom's heights, we will need to update the waitingRooms and their partner attachedRoom if applicable.