Can now move waitingRooms and attachedRooms
When moving the entrance and exits of a puzzleRoom, the waitingRooms would need to adjust their other connected room as waitingRooms and their linked rooms must all share the same height.

Now, the request to move an attachedRoom will come with moving the other attachedRoom it's waitingRoom is connected to.