GlobalRoomController can now delete rooms
By giving a room index, we can now delete a room from the game. When the room gets delete, it will remove the puzzle room along with it's two attachedRoom and the waitingRoom at it's entrance. The function will then update it's arrays and reconnect the puzzleRooms if the room removed was not on the front end.