Can now create rooms at any index
When creating a room at a given index, it will duplicate the room at the index and then add it in the proper position in the heirarchy. This allows the room to be linked with the rest of the rooms and reconnect all the rooms. However, it assumes there is enough space between the room defined by the given index and the room ahead.