Can now move multiple puzzleRooms at once
When moving a puzzleRoom, the user has the choice to move the given index plus any puzzleRooms that follow. This will be helpful when trying to extend one room while keeping the relative distance between the following rooms the same. It will also help when trying to insert a new room into the middle of the list.