Can now create new rooms at any index and adjust the position of the upcomming rooms
When creating a new puzzleRoom, it will duplicate the given index and move the room ahead enough to properly fit into the game. It will also move the upcomming rooms too as it finishes the connecting between the puzzleRooms. The only thing to note is the waitingRoom's sizes are hard coded into the function.

Note there is an issue with waitingRooms when creating a new room. The waitingRoom connects itself to an attachedRoom on the puzzleRoom ahead. This is probably because of something to do with the way we instantiate a new object.