Create a new first puzzleRoom
The new first puzzle room will be used twice. The goal of this is to introduce the player to portals. The first time it appears, they can just walk through. However, the next time the room appears, they will need to hug a wall to find a portal that connects a pillar to a wall, letting them walk on the ceiling.