Created a second new puzzleRoom and an empty room to properly reconnect the rooms
Now, the first 2 puzzle rooms are simple, and the second 2 puzzleRooms are deviants of the first two. The 5th puzzleRoom will be used get the player to properly position themselves so that we can connect the 4th room to the 6th room, as the 3rd and 4th room will flip and position the entrances at different heights.

So far, the 5th puzzleRoom is empty. An idea could be similar to the last puzzleRoom, but a much simpler design (introduce the concept of clear "gates", show throughout the final puzzle room).