Fixed a few bugs, Revamped room 3.
A few bug fixes:

In room 7, the double portal path that lead outside the room did not share the same hieghts.

The windows and pillars while falling were flickering, so a change to the player's render distance has fixed the windows and increasing the density of the flooring with the pillars has mostly fixed that issue.

Also, a new layout for room 3 has been implemented. This new layout should will ensure the player passes through the portals on the side of the room.