Imported sunflare to make the outside seem more lively
Imported the default unity's sunflare object and added it to the main directional light that is used as the sun. The startingRoom's window cannot properly place/show the sun as it is part of the skybox, so certain adjustments to the window will need to be made.

The inclusion of a sunflare makes the diffuse lighting on the terrain seam much more natural now that there is a relative sun in the sky that it clearly is reflecting from.

Also, the StartingRoom's reference circle is removed.