Further fixed and added cubes to the final room.
The player was able to enter a portal in the final room if they were to land on top of one of their edges and run into the portal as they fell. We do not want to player to be able to through a portalin that room in an unintended way.

Also, added a lot of extra cubes around the portals making it less simple than the rest of the rooms.