Adjusted room 7 to make it more clear
The player will now be inclined to walk through the first portal and enter the final room on the wrong angle. Also, when falling into this room at the right angle from the final room, there's a better chance the player won't miss the teleporter to fall into the next waitingRoom.