Adjusted the player's airborn movement speed and added a portal in room 7
The first portal encountered in room 7 now has a portla behind it, making it a 2 way entrance. Also, the airborn movement speed of the player is reduced so that when the re-enter room 7 from the final room with the right angle, it is impossible to avoid the first teleporter they fall into.