Added a basic implementation of a conditional in the shader
The custom unlit shader has a basic implementation of an if statement without using the actual conditional. This is to preserve the performance and implement a conditional.

Also, after testing camera effects there is new a bunch of new built-in shaders in the editor which i do not know how to remove that are in the Assets\Standard Assets folder.