Cube can now create it's own inner edge vertices and the triangles for the X positive face are set.
The cube now creates it's own inner edge vertices using the GetBoxVertice function. The UVs for the edges are not properly set, but the positive X face's triangles are being set and shows how the vertices are in the proper positions and that a function can quickly supply each face with their required triangles.