Added a hack to allow a boolean to determine which face needs to be fully shaded
The booleans used to determine which face needs to only use the second material were broken with the addition of the edges. Now, when the user wants a side to be fully shaded, the edge sizes will be set to fill the face. This is a temporary hack until a fade-in between the materials is found.