Optimized and cleaned up the triangle creation functions
Now when we want to set the triangles of a cube, it will only use one function for both light and dark surfaces. The edges however are not yet rendered as they will use a thrid material linked to the cube. This is because the intended blend effect will need to use a custom shader.